move `join_dependent` and `join_independent` to common
